app_web/requirements.txt
flask
google-cloud-storage==1.44.0
google-cloud-firestore==2.2.0